backend/requirements.txt
django-extensions
django-cors-headers
djangorestframework
psycopg2-binary
dj-rest-auth
djoser
